backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pytest==8.2.0
